common TransitionObserver
